// frontend/src/api/models/listFinancePayablePayablesParams.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */
import type { ListFinancePayablePayablesSourceType } from './listFinancePayablePayablesSourceType'
import type { ListFinancePayablePayablesStatus } from './listFinancePayablePayablesStatus'

export type ListFinancePayablePayablesParams = {
  /**
   * Search term (payable number, supplier name)
   */
  search?: string
  /**
   * Supplier ID
   */
  supplier_id?: string
  /**
   * Status
   */
  status?: ListFinancePayablePayablesStatus
  /**
   * Source type
   */
  source_type?: ListFinancePayablePayablesSourceType
  /**
   * From date (ISO 8601)
   */
  from_date?: string
  /**
   * To date (ISO 8601)
   */
  to_date?: string
  /**
   * Filter overdue only
   */
  overdue?: boolean
  /**
   * Page number
   */
  page?: number
  /**
   * Page size
   * @maximum 100
   */
  page_size?: number
}