// frontend/src/api/models/listFinancePayablePayablesStatus.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type ListFinancePayablePayablesStatus =
  (typeof ListFinancePayablePayablesStatus)[keyof typeof ListFinancePayablePayablesStatus]

export const ListFinancePayablePayablesStatus = {
  PENDING: 'PENDING',
  PARTIAL: 'PARTIAL',
  PAID: 'PAID',
  REVERSED: 'REVERSED',
  CANCELLED: 'CANCELLED',
} as const