// frontend/src/api/models/listFinancePayablePayablesSourceType.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type ListFinancePayablePayablesSourceType =
  (typeof ListFinancePayablePayablesSourceType)[keyof typeof ListFinancePayablePayablesSourceType]

export const ListFinancePayablePayablesSourceType = {
  PURCHASE_ORDER: 'PURCHASE_ORDER',
  PURCHASE_RETURN: 'PURCHASE_RETURN',
  MANUAL: 'MANUAL',
} as const